Authorization code of not active project cannot be used
Authorization codes of apps in a not active project can
not be used anymore. While in the case of the authorization code
it is not that important due to the short validity in time, the same
restriction will be applied to the access tokens and to the future
refresh tokens se everything should be consistent.

Test scenario:
 1. In the project admin create an OAuth2 app (note the ID and the given secret)
 2. Access the authorize page at the URL
https://tuleap.example.com/oauth2/authorize?client_id=<client_d>&scope=demo&response_type=code&redirect_uri=<redirect_uri>
 3. Quickly retrieve (it is valid only 1 minute) the authorization code from the URL
 4. Quickly suspend the project
 4. Try to exchange the authorization code for an access token:
    shell> curl -X POST -H 'Content-Type: application/x-www-form-urlencoded' \
     --user '<client_id>:<client_secret>' \
     --data 'grant_type=authorization_code&redirect_uri=<redirect_uri>&code=<authorization_code>' \
     https://tuleap.example.com/oauth2/token

This is part of story #14542: have OAuth2 flow

Change-Id: I615160eab1c1208d2c87235059ea82406bdd14be