Introduce Monolog for Syslog based logging
syslog is the unix standard for logging stuff, instead of throwing logs in
a gazillion of log files that are a pity to parse/find/correlate syslog
provides a service to centralize everything and do stuff on it (like
forwarding to other services).

How to test it:
- tuleap config-set sys_logger syslog
- journalctl -t tuleap -f

With this first commit, only the logs relative to "BackendLogger::getDefaultLogger"
are converted to make the change small.

Part of story #14683 install a basic image

Change-Id: If3ecac5a897cb398d45cbe9e0944204b3057e052