List authorized OAuth2 app in user settings
Part of story #14543 have OAuth2 user settings

How to test:
- Browse to your account settings. You should see a new tab "OAuth2
Apps"
- When you visit this tab, there should be an empty state.
- Create a new OAuth2 app in a project administration.
- Access the authorization endpoint. '<client_id> and <redirect_uri>
must be replaced. <redirect_uri> should be URL-encoded.

https://tuleap.example.com/oauth2/authorize?client_id=<client_id>&state=xyz&response_type=code&scope=demo%20read:project&redirect_uri=<redirect_uri>

- Authorize the app.
- Visit again the "OAuth2 Apps" tab in your settings. You should see the
App you just authorize and the project in which you registered it.

Notes: Scopes granted to the app will be added in a later contribution.
The "Revoke" button is disabled on purpose and will be handled later.

Change-Id: I129aa4009008e8269295e0e2dfcaeaa65e3301e1