Move OAuth2 access tokens table in the oauth2_server plugin
No functional change is expected from this contribution.
This is a technical change to move the DB tables in the
OAuth2 server plugin.

This change is needed because the access token needs to
know which authorization they are related to. Without moving
all the data we will end up with data duplication and/or far too
many events between the core and the plugin.

This is part of story #14542: have OAuth2 flow

Change-Id: Ie8a0ae8bd7c9387dda859f3c1b7c7afc49eceb41