Add missing index on the plugin_oauth2_authorization table
The app_id column is used to filter the authorizations when deleting
and app. Since there is only a composite index on (user_id, app_id)
MySQL is forced to do a full scan to find the rows. This can be a source
of slowness when deleting an app.

Part of story #14541: have OAuth2 project admin

Change-Id: Ibb6c2856cf0d61d3dc2bd249a6aae06ccb213c5b