request #14701: (un)Restriction of resources for projects is not error prone
Following Bastien & Scapin's ergonomic criteria, restriction of resource
(for exemple allowing all projects to access to a plugin) should display
an advisory message requesting user confirmation when there maybe data
loss (in our exemple the list of previously whitelisted projects).

Furthermore the label "Allow all the projects" of the switch takes all
the width of the page. This means that when user misclick in the middle
of white space, the switch is activated. We should reduce the click zone
to the minimum.

Change-Id: I56934d2afaadb3a03f0779ef1bbde220601b58ad