Save scopes requested for each authorization
For now this information is stored in the database and not yet used.
This information will be used to create access token with the requested
scopes instead of always issuing an access token with the scope 'demo'.

To test:
 1. In the project admin create an OAuth2 app (note the ID and the given secret)
 2. Access the authorize page at the URL
https://tuleap.example.com/oauth2/authorize?client_id=<client_d>&scope=demo&response_type=code&redirect_uri=<redirect_uri>
 3. Accept the requested authorization
 4. The information should be stored in the
plugin_oauth2_authorization_code_scope tables

This is part of story #14542: have OAuth2 flow

Change-Id: Ie7fc73fd7457736f9f0747d02938b641a45c5c98