Project admin can choose if the usage of PKCE [0] is mandatory or not when creating a new OAuth2 app
App that are not forced to use PKCE can still decide to use it
opportunistically if they can but it will not be enforced.
For now, the information is only stored and displayed.

PKCE is recommended when using the authorization code grant [1] and
might become mandatory in future versions of OAuth2. Security
recommendations make the support of PKCE mandatory for authorization
servers.

This is part of story #14542: have OAuth2 flow

[0] https://tools.ietf.org/html/rfc7636
[1] https://tools.ietf.org/html/draft-ietf-oauth-security-topics-14#section-2.1.1

Change-Id: I37e3df3155c2be487344819cdd5ad2e75497ff81