Replace dedicated loggers that extends TruncatedLevelLogger
As always inheritence is creating too much coupling and it's hard to
introduce the new logging mecanism inside existing loggers.

So the design choice is to remove their usage as much as possible and
to replace by static factory methods or at least usage of a const when
changing the getLogger() method would trigger too much refactoring.

The hardest spots are left for dedicated patch.

Part of story #14683 install a basic image

Change-Id: I88fe0ca356fb09541d07af5dafe47a05cd9cd0e4