Initialize dbauthuser grants with mysql-init command
One step closer to completly remove the dependency on mysql CLI
for Tuleap setup so TLS can be enabled with less pain.

I took the opportunity to rename the various arguments to make
them easier to understand.

The NSS part fill me of sadness because, by design it's not possible
to grant permissions on tables that doesn't exists yet without
granting CREATE to user too... so we need to grant both SELECT and
GRANT and revoke CREATE right after.

Part of story #14683 install a basic image

Change-Id: Ie51c5a148e8fbe243c2bdfca67a691594af9231b