Initialize mediawiki database with setup:mysql-init
There is a leak of mediawiki in the core but it was already there
with the shell version of init and solving the leak problem is a
tough one (need to introduce a plugin system to tuleap-cfg).

As there as no longer any mysql cli call in the "remote db" scenario
of the install, I also removed the mysqlcli and mysql_check. the
availability of the mysql server is verified in related commands.

Part of story #14683 install a basic image

Change-Id: Iad06bb33561030782288b98df5058957e7fae4fb