Base of TEE Docker plugin
It's gonna be a weird plugin as it's unlikely to do anything in Tuleap
itself.

As of today, it's purpose is to expose a command to be run
at the start of docker TEE containers. As this cannot be part of
`tuleap` CLI (that requires a working Tuleap) and `tuleap-cfg`
doesn't have plugin support, the less bad tradeoff is a dedicated
CLI tool.

Part of story #14683 install a basic image

Change-Id: I77287424fcd20f65648d8ab9f7f11b5c01e5b24f