request #14721: Some charts cannot be built with gd 2.3.0
This more a workaround than anything else. It is unclear at the moment
if the issue is due to libgd directly or a misuse of libgd in ext-gd.

To test you need to update the gd-last package to the latest version and
restart all the process that might have loaded it.

Change-Id: Iea22a14980708217885f32978ea789e5e9505f41