Move built scripts out of www/
Part of request #14752 Unify scripts and themes build for Core

Everything is supposed to work as before.

Notes:
Once again select2 has given me hell!
I had to remove @types/select2 from TLP and fork it due to a types
conflict between @types/node and @types/requirejs. The latter is a
dependency of @types/select2, although we never use it in production
code. I could not figure out how to exclude @types/requirejs, so the
only solution left was to fork @types/select2 and remove manually the
dependencies to @types/requirejs.
The situation is unlikely to be resolved, see a related PR at [0]

[0]: https://github.com/DefinitelyTyped/DefinitelyTyped/pull/35946

Change-Id: I81b599c0094c13379fbebbe1abfb0d854044436f