Add a read only OAuth2 scope to access tracker information
It is now possible to request the scope 'read:tracker' in the
authorization request to get access to REST API of the trackers (and to
the information they contains) in a read only mode.

The primary goal of the contribution is to introduce the remaining
building blocks to make the introduction of new OAuth2 scopes trivial in
the future.

Part of story #14542: have OAuth2 flow

Change-Id: Ife8ff979a697ae9a9ea7f627eb47930b46846e41