Move the declaration of the REST endpoint GET /projects/:id/trackers in the tracker plugin
The declaration in the core part of the plugin prevent us to use an
OAuth2 scope declared in the tracker plugin for this route. In any
cases, the declaration in the core is not needed has the plugin can
handle the declaration itself.

No functional change.

Part of story #14542: have OAuth2 flow

Change-Id: Icfd28cb12a58a6029236c96da63f49c416bca65f