Dnd or copy/paste images in steps
In artifact view, you can now drag 'n drop or copy/paste images in:

* Text fields
* Step definitions
* Step expected results
* Follow up comments

(as soon as they are in HTML format)

Export the project and import it in a new project. The images referenced
in the step definitions or expected results are the imported ones.

Part of story #13506: Support Images DnD in TTM steps

Change-Id: Ibbc601e367e27331ca0f1d68a4a9946d4118a8cf