Add the 'openid' scope
This scope will be used to determine if it is needed to return an ID
token when returning an access token and if the access to the UserInfo
endpoint is allowed.
For now it does not grant access to anything but it can be requested in
an authorization request.

Part of story #14714: be an OpenID Connect provider

Change-Id: I5514bc751557d7574e6395ec9dda781f885615c0