Introduce config-list command
This first patch put in place the structure and get feedback.

This only focus on the variables that can be set with config-set
but it might be interesting to have the other variables as well.
However it might be a lot of info to display and it's not as easy
to get the help for those.

request #14760 Add config-list command to tuleap CLI

Change-Id: I943122a1850503af3028bc9865d02b31ee3b6576