Fix display of default tracker templates
After a little use of recent changes[0], it appears that the way to
display default tracker templates in the middle of custom templates is
wrong: default trackers are not customizable and should not be in a
custom templates section.

Furthermore Tasks, User Stories, Sprints, or Releases trackers are not
really relevant on their own and should not be promoted as default
templates.

With this commit, when we want to create a tracker we have:

1. A section for custom templates

 This section lists all trackers in projects that have been flagged as
 "Template" by site administrator. On a new install, this list is empty,
 therefore the section is not displayed at all.

2. A section for Tuleap templates

 As of today only Bugs and Activity trackers are provided.

 Those trackers will benefit from dedicated icons, but it will come
 later.

3. A section for advanced users

 This section allows to create a tracker from another project, from xml,
 or from scratch (empty).

[0]: git #tuleap/stable/16de5180e43c7df44b086cec9fdd7ed8b8bfe4de

Part of request #14717: Cleanup of tracker templates
Incidentaly fixes request #14758: Tracker creation doesn't show the right template

Change-Id: I8d44debbc9af82294f0c27636a8eea0dac298e9c