request #14127: Render any Markdown file in the Git repo web UI
Before this contribution, only the READMEs were rendered in the tree
view. Now, when you a open Markdown in the Git repo web UI you see the
rendered file instead of the source. It is however still possible to see
the source of a Markdown file if you really want to.

Change-Id: I7dca29b6296939740aabc42c826d0505a7dc2e2c