Prepare the ground to add [view details] button
In order to be able to add a new button to see the full comment, we must
first rearrange the html and css so that it looks like the mockup.

No functional changes, only styles:
* Header of the test is white on grey (red/green/blue) if test is notrun
(failed/passed/blocked)
* Icons and title are aligned
* Results has its own section, white on dark grey (dark red/green/blue)
* Links (if any) in the comments are white and underlined

Button & modal will be added later.
Max height of the results will be added later.

Part of story #14735: include images in test execution comment box

Change-Id: I4cb4d7fc61cfdb9da8015dee9ff8d45b55969b09