Display details of execution comment in a modal
The comment that has been written for an execution (when marked as
passed/failed/blocked/notrun) is now truncated to not take too much
place.

There is a button that displays the full comment in a modal.

To test it, you might need to use the artifact view of the execution to:

* switch the status of the execution
* enter HTML content in results
* DnD image in this content
* (don't forget to set the time field!)

=> Go back to the TTM UI of the test execution:

* The text is truncated
* The button (always present) displays a modal to reveal the full
comment.

Technical note
--------------

In order to not have an image that breaks the modal layout if it is too
large, there is now a max-width 100% on all img tags. We had to do it
manually in various places like in git. This is now the default in TLP
(BurningParrot only).

Part of story #14735 include images in test execution comment box

Change-Id: Ib11cb53fab4df9caf00ce3bbb6370470b5fc5589