Relative link in a rendered README from a subfolder does not work in the tree view
The link will not be correct in the tree view but will be in the blob
view.
To reproduce have a README.md file in a subfolder of a git repo and
reference a file in the top folder with a relative link (e.g.
`../awesome-alpaca.jpg`).

Part of request #14127: Extend markdown support in the git plugin

Change-Id: Iacbaa3010a853a7a294fa1fe1026cde63be4c727