request #14777 XML export/import should preserve traceability in TTM
When we create a campaign, it is linked to a specific version of test
definitions. This means that two different campaigns may pass a
different version of the same test. This is normal since needs and
software evolves and tests definitions should be adapted over time.

Concretely, when you edit a test definition through the artifact view
(update or add a step for example), then existing campaigns still points
to the previous version(s).

Currently the XML export/import broke this:

1. Create an empty campaign.
2. In the campaign, create a new test with description "Version A"
3. Mark the test as failed
4. Export the project
5. Import it in a new one
6. Go to the test definition **artifact view**
7. Update the description to "Version B"
8. Go to the campaign, select the test

Expected result:
=> The test should display "Version A"

Change-Id: Icca4e04474cc51088df4b3cdc31456470c882b80