The authorization endpoint must support to be called with a HTTP POST request [0]
While this is made mandatory by the specification it is unlikely to
happen in real life. Tuleap marks all its cookie with the SameSite=Lax
attribute so the authentication cookie will not be sent cross-origin
resulting in a weird experience for the user.

To test you can create a page looking like this (add PKCE parameters if
needed):

    <html>
    <body>
    <form action="/oauth2/authorize" method="post">
        <input type="hidden" name="response_type" value="code">
        <input type="hidden" name="client_id" value="<client_id>">
        <input type="hidden" name="redirect_uri" value="<redirect_uri>">
        <input type="hidden" name="scope" value="openid">
        <input type="submit">
     </form>
     </body>
     </html>

Part of story #14714: be an OpenID Connect provider

[0] https://openid.net/specs/openid-connect-core-1_0.html#AuthorizationEndpoint

Change-Id: I1953767c569b4fb50ed3c5fdb4e35e8a1052e905