Add a 'read:user_membership' OAuth2 scope
This scope allows to get access to the user groups the suer is in.
Access to GET /user_groups/:id is also granted by this scope to make
the response of GET /users/:id/memberships usable by an external client.

Part of story #14018: have a central management of users and groups using oauth

Change-Id: I71ed6ec9b7387d0c1072c86d68ffe060fc14bd81