REST endpoints cannot be accessed with an OAuth2 access token
A stub of Psr\EventDispatcher\ EventDispatcherInterface has been added
for Psalm to make sure the issue does not happen again in the future.
The stub has been named with .phpstub extension to help IDEs to not
confuse which file is the actual interface and which one is the stub.

Issue introduced by 80d70b188ffe4fd893927d43945af5266e91ad52.

Part of story #14714: be an OpenID Connect provider

Change-Id: I2cb73e1877456837bcdd3448f508f7aa2ae5fe7a