Support the prompt parameter with the none value [0]
To test, access the authorization endpoint with 'prompt=none'.
If everything can be verified without requiring an interaction you will
be redirected with an auth code otherwise (if you are not logged or if
your consent is needed) you will be redirected with an error.

Part of story #14714: be an OpenID Connect provider

[0] https://openid.net/specs/openid-connect-core-1_0.html#AuthRequest

Change-Id: I524cf8878817d95a8516bc5eb2fd8bbc121d8e29