Support the prompt parameter with the login value [0]
To test, access the authorization endpoint 'prompt=login'. Even if you
are already logged in, you will be redirected to the login page to
authenticate again. Note that you can mix this value with the 'consent'
value to force the user to authenticate and consent to the scopes.

Part of story #14714: be an OpenID Connect provider

[0] https://openid.net/specs/openid-connect-core-1_0.html#AuthRequest

Change-Id: Ie7edb7b1796920d964f486a0ce8c7790904138a1