Consent should always be required when the authorization request ask for the offline scope
The OpenID Core specification [0] says that when using the offline_access
scope the prompt parameter must contains the consent value or other
conditions to process the request are in place. To avoid adding yet
another case of error when the prompt parameter is not correctly set,
requesting the offline_access scope now acts as if you had set
'prompt=consent' to the request.

Part of story #14714: be an OpenID Connect provider

[0] https://openid.net/specs/openid-connect-core-1_0.html#OfflineAccess

Change-Id: I9f832e980d31b0b14a941e17924d640acffd44bf