Remove references to /themes/local
It was a hack introduced when we moved to nginx to allow
site administrators to override site logo. It's now handled
properly (without symlinks...).

Regarding 03-location.conf, I didn' remove entirely the management
of the images with try files (L11-15) because I don't know the part
of tuleap that still rely on the fallaback theme -> common if an
image is missing.

Part of request #14786 Custom organization logos cannot be displayed anymore

Change-Id: Iaabc3a254350caf8987a1447062a0efda28aca36