Add support for Azure specifics for MySQL
MySQL in MSFT Azure has a constraint related to usernames.
They must be defined with '@dbname' at call site. For instance
Given

   GRANT SELECT ON tuleap.users TO alice@'%';

Alice must connect like:

   mysql -hsome-db.mysql.database.azure.com -ualice@some-db...

Notice the 'alice@some-db' at call site...

So we introduce a dedicated switch --azure-suffix to be used when
targeting an install on MSFT Azure (it's not like it's the first
time we have to implement Azure specific stuff anyway).

As the stuff becomes quite tricky (Grants different from conf files)
the generation/modifications of configuration files related to
database is now handled by the init command that is happy to have
a dozen of arguments now.

As it's really messy to get stuff right on setup + DB I decided
to remove the badly named --web-server-ip option of setup as:
- either it's a DB in localhost and your DB shouldn't be reachable
  anyway
- or it's a remote DB and it's unlikely that you want to tight
  your application server to one IP. We are in the Cloud area,
  application servers change so are their IPs.

Sad part of story #14768 have a tuleap enterprise edition docker image production ready

Change-Id: Ia8256d3d6a2268bd884879e7099caf417fdadd30