Tracker should be created with an empty report
part of story #14766 Have an import from jira structure

How to test:
Be sure to have tuleap config-set display_jira_importer 1

Fill form creation data.
Tracker have how an empty report.

We need to be able to import artifact and have the field mapping
to be able to add the summary in columns.
It will be done in a dedicated commit later.

Change-Id: I755ad6cdfdad44160d9ea921efcda8ea1d5eee7f