Test that a user can approve an OAuth2 authorization request
In order to achieve that a new service is spawn with the E2E stack to
play the role of the relying party. For now, this relying party is very
basic as its only role is to redirect the user with a valid
authorization request. The retrieved authorization code is not yet used,
there is no proper validation that the OAuth2/OIDC flow is correct.

To make easy to generate a certificate for the mandatory TLS usage and
have the capability to reuse bits of the PHP ecosystem in the upcoming
contributions, the service uses AmpPHP [0]. That makes possible to
dynamically generate a certificate when the service starts without having
to deal with the setup of more "standard" PHP stack (e.g. nginx/PHP FPM,
Apache/mod_php) since only the PHP CLI is needed.

Due to the nature of the test scenario, it has been needed to allow
Cypress to do cross-origin requests.

Part of story #14714: be an OpenID Connect provider

[0] https://amphp.org/

Change-Id: Ifdf2995f3914644a8405d90df1ebc7130c73d240