First part of artifacts import from Jira
This is part of story #14766 Have an import from jira structure

This patchs adds a very basic artifact import process.
Only the summary value is imported without history.

Full field permissions will be done in a dedicated commit.
Pagination of Jira issue will be handled in a dedicated commit.
Submitter and submitted date wll be handled in a dedicated commit.
Jira link will be added in a dedicated commit.

Change-Id: Ic73ffcb1004789528dd58972d677a8618aeae751