request #14797: Add support for inline style <code> in ckeditor
There isn't any dedicated button for <code>, we are forced to use
"Styles" dropdown. This dropdown is reduced to only a few inline style:

* bold
* italic
* code
* superscript
* subscript

Bold and Italic buttons are kept in the toolbar to ease access.

Superscript and Subscript has been asked because it can be useful for
example in the electronic field to write something like:

> V<sub>IO_Sup</sub> shall be strictly lower than V<sub>Sup</sub>

Underline, Subscript, and Superscript buttons cannot be put on the
toolbar because they are explicitely removed in our version of
ckeditor[0]. Underline is not added to the Styles dropdown because it is
rarely a good idea to underline text on the web.

<sub> et <sup> was lacking style in TLP due to its reset. It is now
fixed.

The list of impacted ckeditor instances is the following:

* artifact view (follow-up comment or HTML text fields)
* artifact modal (angular)
* documents » embedded file editor
* admin » statistics » project over quota » warn administrators
* admin » massmail
* admin » global configuration » homepage headline
* frs licences definition
* MyProject widget » contact modal

[0] https://github.com/ckeditor/ckeditor4-releases/blob/standard/4.11.1/config.js#L31

Change-Id: Iebb91ab775a7f22a0bd9c406b611b7202eaa11db