Test the OAuth2 flow with a basic relying party
This contribution extends the capabilities of the relying party used in
the end-to-end tests. The RP can now exchange the authorization code for
an access token and retrieve information from the user info endpoint.

For now, the ID token is not verified at all so the RP is closer to an
OAuth2 client than an OIDC one.

Part of story #14714: be an OpenID Connect provider

Change-Id: I754b5eab885d9ca3ee0513aca623b74e4a5fed01