request #14798: OIDC client plugin uses BasicAuth when accessing the token endpoint
The credentials are no more sent in the body of the request but via the
Basic HTTP authentication scheme. This is the recommended way by the OAuth2
framework [0] and the only that OAuth2 authorization servers must implement.

No functional change is expected as all OAuth2 authorization servers are
expected to support this.

[0] https://tools.ietf.org/html/rfc6749#section-2.3.1

Change-Id: I895f3046f386d5a68927eaff7f1e27750bdc1da4