Replace Restler's Explorer to generate the API specification file
The file is now generated by the API Explorer plugin directly.
A large section of the code used to do the generation comes from the
Luracast\Restler\Explorer\v2\Explorer class.

This change will allow us to information about which OAuth2 scope is
needed to do some operations. It also sets the ground to generate the
API specification file in the OpenAPI 3.0 format.

Do not forget to redeploy nginx configuration and to clear Restler cache
before testing.

Part of story #14714: be an OpenID Connect provider

Change-Id: I7fe0bb3de4232460e9157bcc2c3bd7d1bc868c7e