Display label for comment box
Placeholder is not natively supported by ckeditor. We need to find (or
write) a plugin in order to be able to have placeholder.

Therefore we propose to have only a label for now.

Bonus: add missing border on top of the footer section.

Part of story #14735: include images in test execution comment

Change-Id: I862c63bd970fdea8ded3e68db3434e35b66c0297