Deployment of FPM and redis configuration
The site-deploy:fpm command now allows to configure the
connection to a redis server.

I'm not 100% fan of this design choice but I decided to include
the generation of tuleap `redis.inc` file by the site-deploy:fpm
command for 2 reasons:
- I doesn't really make sense to configure redis for Tuleap but not
  for the php sessions
- Having 2 different commands for that would be too complex for
  deployment.

The generic site-deploy command now handles the fpm/redis generation
in addition to images.

Finally a docker-compose.yml is added as an example and test purpose.
It's based on public enalean/tuleap-aio:centos7 image because it was
good enough for most tests and relying on TEE image would have been
too heavy.

Part of story #14768 have a tuleap enterprise edition docker image production ready

Change-Id: I1d656c2f281b7afdf1ede274a7dc27c61810c678