Initiate cypress tests for TTM
This only bootstraps e2e tests in testmanagement plugin.

It covers the project creation based on AgileALM templates which was
missing.
Now project creation test in core creates a project based on Issue
template, instead of the first random item in a list.

Testing campaigns and tests will be done in dedicated commits.

Part of request #14809: Have cypress tests for TTM

Change-Id: I52f55e01f05919ef477da55eb1c26d1d92b94496