request #14783: Activate by default the new project creation screen
Because it's stable enough.

Try to create a project: the new creation UI is shown.
There is a link to switch back to the legacy UI.
On the legacy UI there is an alert to warn the user that the new UI
should be used instead.

This is the same UX than for tracker creation.

The config option is removed from forge_config table.

Please note that a change has been made in tracker creation in order to
prevent the underline of whitespace between icon and text of "Switch
back" link. This change forces us to prettier-ignore the html element
because prettier wants to insert unwanted whitespace.

Change-Id: Ia88a42de3c44f7f0592a4d583162c84178071809