A user without a password cannot set one herself
As of today this situation can happen when a user has been automatically
registered after a sign in with an OIDC provider marked as being the only
authentication method. With upcoming contributions it will also be the
case for any users that have registered after signing in on any OIDC
provider proposed by the instance.

In order to make sure there was no magic trick (e.g. a null casted
silently to an empty string) the plaintext user password has been
changed for a Tuleap\Cryptography\ConcealedString instance. It makes
easier for the static analysis to catch misuses and anyway it limits the
chance of leaking a user's password.

To test, you can the password field of one of your user to NULL. When
you are authenticated with this user, the password change form should be
gone. If you try to authenticate with this user the authentication must
be denied.

Part of request #14812: Users signing in via an OIDC provider should be
able to create an account without setting a password

Change-Id: I4522f126be60b314c0dbb780fe5491ed64bda7ad