bugs #14829 No errors when a project is displaying without agile dashboard
In the past, we can add only one widget on an project dashboard and display this project in the widget.
Now, we can display any projects on any dashboard (personal or project), so it needs to check
if the displayed project has AD or not.

There are 2 cases to reproduce this bug:

How to reproduce:
1)
	- Have an agile dashboard configuredi in a project
	- Disable service agile dashboard in this project
	- Add project milestones with this project

	=> Project Milestones doesn't throw any errors and display releases
		=> All links to go to AD don't work

2)
	- Have project without service agile dashboard and no tracker
	- Have project milestones

	=> Error 'No Root Planning defined' is not true.

Expected new behaviour with this patch:
	- Now, if service AD is not enabled in the displayed project, there is an error message "Service AgileDashboard is disabled"

Change-Id: I8369b5aebed914fe3be03cad0d94232cdbfb8778