request #14835: Line number does not line up with the line itself
There is a slight misalignment between the line number and the line
itself the Git web UI which accumulates over lines. After ~10 lines the
line number is no more aligned at all.

The issue is caused by the fact that a different font is used to
display the line number and the code itself. The font used for the code
is pushed by the stylesheet of PrismJS (the JS library used to do the
syntax hightlighting) so in order to solve the issue the same font
family is now used in both cases. To limit surprises with future
upgrades of PrismJS, we make sure to overwrite the font family it
defines.
As a base, the current font family defined by PrismJS is now used.

Change-Id: Ifc389251ef97362ff74e6ec58afcc7b680c17fda