Can link to an existing user account by signing in to another OIDC provider
To make sure the linking process cannot be hijacked by a malicious user,
the process is now bound to the user that has start it. This is done by
storing and retrieving the link ID into the user session.

Part of request #14812 Users signing in via an OIDC provider should be able
to create an account without setting a password

Change-Id: I805a1fe9e3c8cf00f626895930691b02c890de0b