request #14843: Add a "copy to clipboard" element when creating a SVN token, an access key or an OAuth2 client secret
This element is created via a Custom Element [0] to make possible to
reuse the behavior in different contexts while keeping it isolated from
the rest of the page.

As IE11 does not support the custom elements, a polyfill is needed. To
not penalize modern parts of Tuleap that do support IE11 (hopefully the
majority of the new features) the polyfill must be imported manually by
the users of the element.

Jest now uses JSDOM 16 to have the support of the Custom Elements API.

[0] https://html.spec.whatwg.org/multipage/custom-elements.html

Change-Id: I584eec01a41f6ffcbc465e9dc821ebbb659de544