Rotate ID token signing keys periodically
Rotating the signing key periodically limits the damage in case of a key
compromise.
The period has been set to a short period of time (1H) to make sure the
Relying-Party know how to handle the key rotation as it forces them to
automate the retrieval of the JWKS document. Without that a Relying-Party
might be tempted to hardcode the signing key which will break the day a
key rotation is really needed.

Also to make sure we play nice with all Relying-Party implementations a
`kid` header [0] has been to the ID token.

Part of story #14714: be an OpenID Connect provider

[0] https://tools.ietf.org/html/rfc7515#section-4.1.4

Change-Id: I7eedb367ca1bef3d73160a89baf7647134778dfa