Perl code cannot verify TLS certificate
Perl code seems to have troubles dealing with TLS connection
that requires verification of the CA. It works on recent fedora
(tested on F31) but not on centos and seems related to some
version of openssl.

Instead of introducing yet another option that people can shoot
themselves in their feets, the cert verification is removed
entirely.

Part of story #14768 have a tuleap enterprise edition docker image production ready

Change-Id: I27b51f35ae438b348db5662cc76b9561e58c07ef